Melhorar Produtos
Principais correções:

 Tipos: int → Long para IDs em todas as classes

DAO: Queries completas com categoria e descricao

 Service: Validações + compatibilidade com Repository

 Entity: Usa Long idProduto consistentemente

Problemas resolvidos:

Incompatibilidade de tipos entre camadas

Campos faltantes nas queries SQL

Métodos incompletos no DAO

Sistema agora compatível e funcional para CRUD completo.

+++ src/main/java/com/tecdes/pedido/model/entity/Produto.java
@@ -2,17 +2,17 @@ package com.tecdes.pedido.model.entity;
 
 public class Produto {
 
-    private int idProduto;
+    private Long idProduto;
     private String nome;
     private double preco;
     private String categoria;
     private String descricao;
 
-    public int getIdProduto() {
+    public Long getIdProduto() {
         return idProduto;
     }
 
-    public void setIdProduto(int idProduto) {
+    public void setIdProduto(Long idProduto) {
         this.idProduto = idProduto;
     }
 
@@ -48,63 +48,3 @@ public class Produto {
         this.descricao = descricao;
     }
 }
-// package com.tecdes.pedido.model.entity;
-
-// public class Produto {
-
-// private Long idProduto;
-// private String nome;
-// private String descricao;
-// private double preco;
-// private String categoria;
-
-// public Produto() {
-// }
-
-// public Produto(String nome, String descricao, double preco, String categoria)
-// {
-// this.nome = nome;
-// this.descricao = descricao;
-// this.preco = preco;
-// this.categoria = categoria;
-// }
-
-// public Long getIdProduto() {
-// return idProduto;
-// }
-
-// public void setIdProduto(Long idProduto) {
-// this.idProduto = idProduto;
-// }
-
-// public String getNome() {
-// return nome;
-// }
-
-// public void setNome(String nome) {
-// if (nome == null || nome.trim().isEmpty()) {
-// throw new IllegalArgumentException("Nome do produto é obrigatório.");
-// }
-// this.nome = nome;
-// }
-
-// public String getDescricao() {
-// return descricao;
-// }
-
-// public void setDescricao(String descricao) {
-// this.descricao = descricao;
-// }
-
-// public double getPreco() {
-// return preco;
-// }
-
-// public void setPreco(double preco) {
-// if (preco <= 0) {
-// throw new IllegalArgumentException("Preço deve ser positivo.");
-// }
-// this.preco = preco;
-// }
-
-// }
